feat: 完成空调调度和账单相关

+++ server/src/main/java/com/hotel/server/serviceImpl/RoomServiceImpl.java
@@ -8,6 +8,8 @@ import com.hotel.common.dto.request.PageRoomReq;
 import com.hotel.common.dto.response.RoomInfoResp;
 import com.hotel.common.entity.Customer;
 import com.hotel.common.service.customer.CustomerService;
+import com.hotel.common.service.server.BillService;
+import com.hotel.common.service.server.CoolService;
 import com.hotel.common.service.timer.TimerService;
 import com.hotel.server.config.IndoorTemperatureConfig;
 import com.hotel.server.mapper.RoomMapper;
@@ -45,6 +47,12 @@ public class RoomServiceImpl extends ServiceImpl<RoomMapper, Room> implements Ro
 
     @DubboReference(check = false)
     private TimerService timerService;
+
+    @DubboReference(check = false)
+    private CoolService coolService;
+
+    @DubboReference(check = false)
+    private BillService billService;
 
     @Resource
     private IndoorTemperatureConfig indoorTemperatureConfig;
@@ -105,12 +113,16 @@ public class RoomServiceImpl extends ServiceImpl<RoomMapper, Room> implements Ro
     @Override
     @Transactional(rollbackFor = Exception.class)
     public Boolean leave(Long roomId, Long customerId) {
-        // todo 先退房, 成功后释放用户占用的所有资源 餐饮 保洁 纳凉 押金
-        // 退房后生成账单并写入 mongo
+        // 先释放用户占用的所有资源 餐饮(这个逻辑改了应该不需要释放了) 纳凉
+        // 然后将用户账单写入数据库
+        // 最后删除房间和用户
+        coolService.turnOff(String.valueOf(customerId));
+        if (!billService.saveBillStatement(String.valueOf(customerId))) {
+            throw new RuntimeException("写入账单失败, 请稍后重试");
+        }
         if (!this.removeById(roomId) || !customerService.removeById(customerId)) {
             throw new RuntimeException("退房失败, 请稍后重试");
         }
-        // 释放资源
         return true;
     }
 }
